+++ db_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/surrealdb/surrealdb.go/pkg/connection"
 	"github.com/surrealdb/surrealdb.go/pkg/constants"
 	"github.com/surrealdb/surrealdb.go/pkg/logger"
+	"github.com/surrealdb/surrealdb.go/pkg/models"
 	"io"
 	rawslog "log/slog"
 	"os"
@@ -153,7 +154,7 @@ func (s *SurrealDBTestSuite) SetupSuite() {
 // Sign with the root user
 // Can be used with any user
 func signin(s *SurrealDBTestSuite) interface{} {
-	authData := &connection.Auth{
+	authData := &models.Auth{
 		Username: "root",
 		Password: "root",
 	}
