+++ pkg/connection/model.go
@@ -56,12 +56,3 @@ var (
 	Patch        RPCFunction = "patch"
 	Delete       RPCFunction = "delete"
 )
-
-// Auth is a struct that holds surrealdb auth data for login.
-type Auth struct {
-	Namespace string `json:"NS,omitempty"`
-	Database  string `json:"DB,omitempty"`
-	Scope     string `json:"SC,omitempty"`
-	Username  string `json:"user,omitempty"`
-	Password  string `json:"pass,omitempty"`
-}
