+++ pkg/models/cbor_test.go
@@ -1,4 +1,4 @@
-package types
+package models
 
 import (
 	"fmt"
