+++ pkg/models/models.go
@@ -1,0 +1,10 @@
+package models
+
+// Auth is a struct that holds surrealdb auth data for login.
+type Auth struct {
+	Namespace string `json:"NS,omitempty"`
+	Database  string `json:"DB,omitempty"`
+	Scope     string `json:"SC,omitempty"`
+	Username  string `json:"user,omitempty"`
+	Password  string `json:"pass,omitempty"`
+}
