+++ pkg/models/replacer.go
@@ -1,4 +1,4 @@
-package types
+package models
 
 import (
 	"reflect"
